src/LangChainAPI/requirements.txt
fastapi==0.100.1
uvicorn==0.23.2
azure-identity==1.13.0
azure-keyvault-secrets==4.7.0
langchain==0.0.256
langchain-experimental==0.0.8
openai==0.27.8
pandas==2.0.3
tabulate==0.9.0